backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
google-genai
python-dotenv
edge-tts
scenedetect[opencv]
openai-whisper
setuptools
crewai
langchain-google-genai
lockfile
